News/requirements.txt
Flask
newsapi-python
nltk
scikit-learn
textblob
python-dotenv
flask_cors
wordcloud 
matplotlib